requirements.txt
django==2.2
cymysql
django-cymysql==2.2
djangorestframework
